@@ -1,40 +1,75 @@
 {
-  "name": "grass-shader",
+  "name": "threejs-mobilehouse-configurator",
   "version": "1.0.0",
-  "description": "Grass shader using simplex noise.",
+  "description": "ThreeJS-MobileHouse-Configurator",
   "keywords": [
-    "noise",
-    "shader",
-    "simplex",
-    "grass"
+    "ThreeJS",
+    "MobileHouse",
+    "Configurator"
   ],
   "main": "src/index.js",
   "dependencies": {
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "@babel/core": "^7.24.4",
+    "@babel/plugin-transform-private-property-in-object": "^7.24.1",
     "@emotion/react": "latest",
     "@emotion/styled": "latest",
     "@mui/icons-material": "^5.0.0",
+    "@types/three": "^0.151.0",
     "@mui/material": "latest",
-    "@react-three/drei": "4.1.8",
-    "@react-three/fiber": "6.0.14",
+    "@react-three/drei": "9.105.4",
+    "@react-three/fiber": "^8.0.8",
     "@react-three/flex": "^0.7.0",
     "@react-three/lightmap": "0.0.2",
-    "@react-three/postprocessing": "2.0.5",
+    "@react-three/postprocessing": "2.6.1",
     "babel-plugin-glsl": "^1.0.0",
+    "babel-preset-react-app": "^10.0.1",
     "glsl-curl-noise2": "0.0.4",
     "leva": "0.9.13",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "postprocessing": "^6.35.3",
     "react-scripts": "4.0.3",
     "react-three-editable": "^0.8.1",
     "react-three-editor": "^2.2.1",
     "react-three-flex": "^0.5.2",
     "simplex-noise": "2.4.0",
-    "three": "0.127.0",
-    "three-mesh-bvh": "^0.4.3",
-    "valtio": "^1.2.3"
+    "valtio": "^1.2.3",
+    "@babel/runtime": "^7.11.2",
+    "@mediapipe/tasks-vision": "0.10.8",
+    "@monogrid/gainmap-js": "^3.0.5",
+    "@react-spring/three": "~9.6.1",
+    "@use-gesture/react": "^10.2.24",
+    "camera-controls": "^2.4.2",
+    "cross-env": "^7.0.3",
+    "detect-gpu": "^5.0.28",
+    "glsl-noise": "^0.0.0",
+    "maath": "^0.10.7",
+    "meshline": "^3.1.6",
+    "react-composer": "^5.0.3",
+    "stats-gl": "^2.0.0",
+    "stats.js": "^0.17.0",
+    "suspend-react": "^0.1.3",
+    "three": "^0.151.0",
+    "three-mesh-bvh": "^0.7.0",
+    "three-stdlib": "^2.29.4",
+    "troika-three-text": "^0.49.0",
+    "tunnel-rat": "^0.1.2",
+    "utility-types": "^3.10.0",
+    "uuid": "^9.0.1",
+    "zustand": "^3.7.1",
+    "hls.js": "1.3.5"
   },
   "devDependencies": {
-    "typescript": "3.8.3"
+    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
+
+  },
+  "peerDependencies": {
+    "@react-three/fiber": ">=8.0",
+    "react": ">=18.0",
+    "react-dom": ">=18.0",
+    "three": ">=0.137"
   },
   "scripts": {
     "start": "react-scripts start",
