@@ -9,50 +9,451 @@ export default function Model(props) {
   const group = useRef()
   const { nodes, materials } = useGLTF('/Mobilehouse_inside.glb')
   return (
-    <group ref={group} {...props} dispose={null}>
+    <group {...props} dispose={null}>
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall.geometry}
+        material={materials.WoodenFloor}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall001.geometry}
+        material={materials.WoodenFloor}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall002.geometry}
+        material={materials.WoodenFloor}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall003.geometry}
+        material={materials.WoodenFloor}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Handle_Argos003.geometry}
+        material={materials['aluminium used.001']}
+        position={[0.487, 0.056, -0.086]}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Handle_Argos004.geometry}
+        material={materials['aluminium used.002']}
+        position={[0.87, 0.056, -1.868]}
+        rotation={[0, Math.PI / 2, 0]}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall004.geometry}
+        material={materials.WoodenFloor}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall005.geometry}
+        material={materials.WoodenFloor}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall006.geometry}
+        material={materials.Tiles}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.topback.geometry}
+        material={materials['Artificial Marble 1']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.top.geometry}
+        material={materials['Artificial Marble 1']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.simple_tap01001.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plate_Stack001.geometry}
+        material={materials.Cup}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane002.geometry}
+        material={materials['cooking mat']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.hood.geometry}
+        material={materials['brushed metal hood']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle013.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle012.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle011.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle010.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle009.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle008.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle007.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle005.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle004.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle003.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle002.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handle001.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes['Cup-Large-White003'].geometry}
+        material={materials.Cup}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube004.geometry}
+        material={materials.plastic}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube002.geometry}
+        material={materials['brushed metal hood']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cooking_stove.geometry}
+        material={materials['brushed metal']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle005.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle002.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_wall_40004.geometry}
+        material={materials['cabinets gray']}
+        position={[-1.698, 0.702, 1.986]}
+        rotation={[0, Math.PI / 2, 0]}
+        scale={[1.518, 1, 1.433]}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_wall_40003.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_wall_40002.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_wall_40001.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_60x82003.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_60x82002.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_60x82001.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82016.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82015.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82014.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82013.geometry}
+        material={materials['front pannels']}
+        position={[-1.241, 1.062, 1.973]}
+        rotation={[0, Math.PI / 2, 0]}
+        scale={[1.518, 1, 1.433]}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82012.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82011.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82010.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82009.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82008.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82006.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82005.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82004.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82003.geometry}
+        material={materials['front pannels']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.bootom_stand.geometry}
+        material={materials['cabinets gray']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.back_decor_wall.geometry}
+        material={materials['back wall']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_wall_40005.geometry}
+        material={materials['cabinets gray.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.cabinet_ground_40x82017.geometry}
+        material={materials['cabinets gray.002']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.houseWall007.geometry}
+        material={materials['Tiles.001']}
+      />
+      <mesh castShadow receiveShadow geometry={nodes.tap.geometry} material={materials.Chrome} />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.sink_hole.geometry}
+        material={materials['Anodized.001']}
+      />
+      <mesh castShadow receiveShadow geometry={nodes.sink.geometry} material={materials.Ceramic} />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.right_upper_door.geometry}
+        material={materials['White - Cristallo']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handles.geometry}
+        material={materials['Anodized.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle003.geometry}
+        material={materials.Chrome}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.handles001.geometry}
+        material={materials['Anodized.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube002_1.geometry}
+        material={materials['WinDoor_Glass.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube002_2.geometry}
+        material={materials['Edit_WinDoors.001']}
+      />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Door_PI002.geometry}
-        material={nodes.Door_PI002.material}
+        material={materials['Wood Base 01']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Frame.geometry}
-        material={nodes.Frame.material}
+        material={materials['Wood Base 01']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Door_PI002001.geometry}
+        material={materials['Wood Base 01']}
+        position={[0.484, 0.173, -1.015]}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Frame001.geometry}
+        material={materials['Wood Base 01']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Handle_Argos001.geometry}
         material={materials['aluminium used']}
-        position={[0.48, 0.06, -1.21]}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Door_PI002001.geometry}
-        material={nodes.Door_PI002001.material}
-        position={[0.48, 0.17, -1.02]}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Frame001.geometry}
-        material={nodes.Frame001.material}
+        position={[0.479, 0.056, -1.207]}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Door_PI002002.geometry}
-        material={nodes.Door_PI002002.material}
+        material={materials['Wood Base 01']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Frame002.geometry}
-        material={nodes.Frame002.material}
+        material={materials['Wood Base 01']}
       />
       <mesh
         castShadow
@@ -63,44 +464,130 @@ export default function Model(props) {
       <mesh
         castShadow
         receiveShadow
+        geometry={nodes.Cube018.geometry}
+        material={materials['grid plastic']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube018_1.geometry}
+        material={materials['darkish grey plastic']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube018_2.geometry}
+        material={materials.screen}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane001_1.geometry}
+        material={materials.sink}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane001_2.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane001_3.geometry}
+        material={materials['sink side hole']}
+      />
+      <mesh castShadow receiveShadow geometry={nodes.Plane_1.geometry} material={materials.grate} />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane_2.geometry}
+        material={materials['brushed metal']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane_3.geometry}
+        material={materials.plastic}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.NurbsPath001_1.geometry}
+        material={materials.Anodized}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.NurbsPath001_2.geometry}
+        material={materials['tap hole']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.prop_chair_karlstad.geometry}
+        material={materials['Chair fabric']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.prop_chair_karlstad_1.geometry}
+        material={materials['chair wood']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane010.geometry}
+        material={materials['metal.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane010_1.geometry}
+        material={materials.wood}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane010_2.geometry}
+        material={materials['plastic.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
         geometry={nodes.back002.geometry}
-        material={nodes.back002.material}
+        material={materials['Dark glossy']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.decorative_board002.geometry}
-        material={nodes.decorative_board002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.shelves.geometry}
-        material={nodes.shelves.material}
+        material={materials['Dark glossy']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.handles002.geometry}
-        material={nodes.handles002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.legs002.geometry}
-        material={nodes.legs002.material}
+        material={materials.Metal}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.left_drawersdoor.geometry}
-        material={nodes.left_drawersdoor.material}
+        material={materials['Wood.001']}
       />
+      <mesh castShadow receiveShadow geometry={nodes.legs002.geometry} material={materials.Metal} />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.right_door001.geometry}
-        material={nodes.right_door001.material}
+        material={materials['Wood.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.shelves.geometry}
+        material={materials['Dark glossy']}
       />
       <mesh
         castShadow
@@ -112,25 +599,67 @@ export default function Model(props) {
         castShadow
         receiveShadow
         geometry={nodes.Cube053_1.geometry}
-        material={nodes.Cube053_1.material}
+        material={materials['Dark glossy']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cylinder032.geometry}
+        material={materials.Metallight}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cylinder032_1.geometry}
+        material={materials['LED-Lighitngs']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cylinder015.geometry}
+        material={materials.Metallight}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cylinder015_1.geometry}
+        material={materials['LED-Lighitngs']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes['Toilet-01_1'].geometry}
+        material={materials['white ceramic']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes['Toilet-01_2'].geometry}
+        material={materials['chrome metal']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes['Toilet-01_3'].geometry}
+        material={materials['White plastic']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Cube000.geometry}
-        material={nodes.Cube000.material}
+        material={materials['porcelain white']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Cube001.geometry}
-        material={nodes.Cube001.material}
+        material={materials['porcelain white']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.sink_1.geometry}
-        material={nodes.sink_1.material}
+        material={materials['porcelain white']}
       />
       <mesh
         castShadow
@@ -184,62 +713,66 @@ export default function Model(props) {
         castShadow
         receiveShadow
         geometry={nodes.Cube007.geometry}
-        material={nodes.Cube007.material}
+        material={materials['Scuffed Aluminum']}
       />
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Door_left.geometry}
-        material={nodes.Door_left.material}>
+        material={materials['Scuffed Aluminum']}>
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Door_left_glass.geometry}
-          material={nodes.Door_left_glass.material}
+          material={materials['Architectural Glass']}
         />
       </mesh>
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Door_right.geometry}
-        material={nodes.Door_right.material}>
+        material={materials['Scuffed Aluminum']}>
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Door_right_glass.geometry}
-          material={nodes.Door_right_glass.material}
+          material={materials['Architectural Glass']}
         />
       </mesh>
-      <mesh castShadow receiveShadow geometry={nodes.DOWN.geometry} material={nodes.DOWN.material}>
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.DOWN.geometry}
+        material={materials['Material.002']}>
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Body.geometry}
-          material={nodes.Body.material}
+          material={materials['Scuffed Aluminum']}
         />
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Wall.geometry}
-          material={nodes.Wall.material}
+          material={materials['Scuffed Aluminum']}
         />
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Wall_glass.geometry}
-          material={nodes.Wall_glass.material}
+          material={materials['Architectural Glass']}
         />
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Wall_right.geometry}
-          material={nodes.Wall_right.material}
+          material={materials['Scuffed Aluminum']}
         />
         <mesh
           castShadow
           receiveShadow
           geometry={nodes.Wall_right_glass.geometry}
-          material={nodes.Wall_right_glass.material}
+          material={materials['Architectural Glass']}
         />
         <mesh
           castShadow
@@ -251,7 +784,7 @@ export default function Model(props) {
           castShadow
           receiveShadow
           geometry={nodes.Plane028_1.geometry}
-          material={nodes.Plane028_1.material}
+          material={materials['Material.002']}
         />
         <mesh
           castShadow
@@ -263,14 +796,74 @@ export default function Model(props) {
           castShadow
           receiveShadow
           geometry={nodes.Plane028_3.geometry}
-          material={nodes.Plane028_3.material}
+          material={materials['Scuffed Aluminum']}
         />
       </mesh>
       <mesh
         castShadow
         receiveShadow
         geometry={nodes.Grid.geometry}
-        material={nodes.Grid.material}
+        material={materials['Scuffed Aluminum']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube073.geometry}
+        material={materials['White - Cristallo']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube073_1.geometry}
+        material={materials.Chocofur_Wood_Solid_01}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle002_1.geometry}
+        material={materials['Anodized.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle002_2.geometry}
+        material={materials.mesh}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube071.geometry}
+        material={materials['White - Cristallo']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Cube071_1.geometry}
+        material={materials.Chocofur_Wood_Solid_01}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane030.geometry}
+        material={materials.mirror}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Plane030_1.geometry}
+        material={materials['White - Cristallo']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle001.geometry}
+        material={materials['Anodized.001']}
+      />
+      <mesh
+        castShadow
+        receiveShadow
+        geometry={nodes.Circle001_1.geometry}
+        material={materials.mesh}
       />
       <mesh
         castShadow
@@ -301,600 +894,6 @@ export default function Model(props) {
         receiveShadow
         geometry={nodes.Chocofur_Free_Beds_01002_4.geometry}
         material={materials['Unmade bed.004']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube002_1.geometry}
-        material={materials['WinDoor_Glass.001']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube002_2.geometry}
-        material={materials['Edit_WinDoors.001']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Handle_Argos003.geometry}
-        material={materials['aluminium used.001']}
-        position={[0.49, 0.06, -0.09]}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Handle_Argos004.geometry}
-        material={materials['aluminium used.002']}
-        position={[0.87, 0.06, -1.87]}
-        rotation={[0, Math.PI / 2, 0]}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall006.geometry}
-        material={materials.Tiles}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube018.geometry}
-        material={materials['grid plastic']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube018_1.geometry}
-        material={materials['darkish grey plastic']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube018_2.geometry}
-        material={materials.screen}
-      />
-      <mesh castShadow receiveShadow geometry={nodes.Plane001.geometry} material={materials.sink} />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane001_1.geometry}
-        material={nodes.Plane001_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane001_2.geometry}
-        material={materials['sink side hole']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.simple_tap01001.geometry}
-        material={nodes.simple_tap01001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plate_Stack001.geometry}
-        material={nodes.Plate_Stack001.material}
-      />
-      <mesh castShadow receiveShadow geometry={nodes.Plane_1.geometry} material={materials.grate} />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane_2.geometry}
-        material={nodes.Plane_2.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane_3.geometry}
-        material={nodes.Plane_3.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.NurbsPath001_1.geometry}
-        material={nodes.NurbsPath001_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.NurbsPath001_2.geometry}
-        material={materials['tap hole']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.hood.geometry}
-        material={nodes.hood.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle013.geometry}
-        material={nodes.handle013.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle012.geometry}
-        material={nodes.handle012.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle011.geometry}
-        material={nodes.handle011.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle010.geometry}
-        material={nodes.handle010.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle009.geometry}
-        material={nodes.handle009.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle008.geometry}
-        material={nodes.handle008.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle007.geometry}
-        material={nodes.handle007.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle005.geometry}
-        material={nodes.handle005.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle004.geometry}
-        material={nodes.handle004.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle003.geometry}
-        material={nodes.handle003.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle002.geometry}
-        material={nodes.handle002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handle001.geometry}
-        material={nodes.handle001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes['Cup-Large-White003'].geometry}
-        material={nodes['Cup-Large-White003'].material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube004.geometry}
-        material={nodes.Cube004.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube044.geometry}
-        material={materials['Roughness Transparent Plastic']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube044_1.geometry}
-        material={materials.light}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube002.geometry}
-        material={nodes.Cube002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cooking_stove.geometry}
-        material={nodes.cooking_stove.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle005.geometry}
-        material={nodes.Circle005.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle002.geometry}
-        material={nodes.Circle002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_wall_40004.geometry}
-        material={nodes.cabinet_wall_40004.material}
-        position={[-1.7, 0.7, 1.99]}
-        rotation={[0, Math.PI / 2, 0]}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_wall_40003.geometry}
-        material={nodes.cabinet_wall_40003.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_wall_40002.geometry}
-        material={nodes.cabinet_wall_40002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_wall_40001.geometry}
-        material={nodes.cabinet_wall_40001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_60x82003.geometry}
-        material={nodes.cabinet_ground_60x82003.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_60x82002.geometry}
-        material={nodes.cabinet_ground_60x82002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_60x82001.geometry}
-        material={nodes.cabinet_ground_60x82001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82016.geometry}
-        material={nodes.cabinet_ground_40x82016.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82015.geometry}
-        material={nodes.cabinet_ground_40x82015.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82014.geometry}
-        material={nodes.cabinet_ground_40x82014.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82013.geometry}
-        material={nodes.cabinet_ground_40x82013.material}
-        position={[-1.24, 1.06, 1.97]}
-        rotation={[0, Math.PI / 2, 0]}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82012.geometry}
-        material={nodes.cabinet_ground_40x82012.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82011.geometry}
-        material={nodes.cabinet_ground_40x82011.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82010.geometry}
-        material={nodes.cabinet_ground_40x82010.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82009.geometry}
-        material={nodes.cabinet_ground_40x82009.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82008.geometry}
-        material={nodes.cabinet_ground_40x82008.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82006.geometry}
-        material={nodes.cabinet_ground_40x82006.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82005.geometry}
-        material={nodes.cabinet_ground_40x82005.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82004.geometry}
-        material={nodes.cabinet_ground_40x82004.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82003.geometry}
-        material={nodes.cabinet_ground_40x82003.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.bootom_stand.geometry}
-        material={nodes.bootom_stand.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_wall_40005.geometry}
-        material={materials['cabinets gray.001']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.cabinet_ground_40x82017.geometry}
-        material={materials['cabinets gray.002']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cylinder032.geometry}
-        material={nodes.Cylinder032.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cylinder032_1.geometry}
-        material={nodes.Cylinder032_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cylinder015.geometry}
-        material={nodes.Cylinder015.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cylinder015_1.geometry}
-        material={nodes.Cylinder015_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes['Toilet-01_1'].geometry}
-        material={materials['white ceramic']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes['Toilet-01_2'].geometry}
-        material={materials['chrome metal']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes['Toilet-01_3'].geometry}
-        material={materials['White plastic']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall007.geometry}
-        material={materials['Tiles.001']}
-      />
-      <mesh castShadow receiveShadow geometry={nodes.tap.geometry} material={nodes.tap.material} />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.sink_hole.geometry}
-        material={nodes.sink_hole.material}
-      />
-      <mesh castShadow receiveShadow geometry={nodes.sink.geometry} material={materials.Ceramic} />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handles.geometry}
-        material={nodes.handles.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle003.geometry}
-        material={nodes.Circle003.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.handles001.geometry}
-        material={nodes.handles001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.back_decor_wall.geometry}
-        material={materials['back wall']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.topback.geometry}
-        material={nodes.topback.material}
-      />
-      <mesh castShadow receiveShadow geometry={nodes.top.geometry} material={nodes.top.material} />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.prop_chair_karlstad.geometry}
-        material={materials['Chair fabric']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.prop_chair_karlstad_1.geometry}
-        material={materials['chair wood']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane002.geometry}
-        material={materials['cooking mat']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.right_upper_door.geometry}
-        material={nodes.right_upper_door.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane030.geometry}
-        material={materials.mirror}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane030_1.geometry}
-        material={nodes.Plane030_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane010.geometry}
-        material={materials['metal.001']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane010_1.geometry}
-        material={materials.wood}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Plane010_2.geometry}
-        material={materials['plastic.001']}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle002_1.geometry}
-        material={nodes.Circle002_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle002_2.geometry}
-        material={nodes.Circle002_2.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle001.geometry}
-        material={nodes.Circle001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Circle001_1.geometry}
-        material={nodes.Circle001_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall.geometry}
-        material={nodes.houseWall.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall001.geometry}
-        material={nodes.houseWall001.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall002.geometry}
-        material={nodes.houseWall002.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall003.geometry}
-        material={nodes.houseWall003.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall004.geometry}
-        material={nodes.houseWall004.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.houseWall005.geometry}
-        material={nodes.houseWall005.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube073.geometry}
-        material={nodes.Cube073.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube073_1.geometry}
-        material={nodes.Cube073_1.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube071.geometry}
-        material={nodes.Cube071.material}
-      />
-      <mesh
-        castShadow
-        receiveShadow
-        geometry={nodes.Cube071_1.geometry}
-        material={nodes.Cube071_1.material}
       />
     </group>
   )
